move lockmeter to metrics package
